feat(): create all entities

+++ src/main/java/es/ediae/master/programacion/gestionusuario/dtos/UsuarioGetDTO.java
@@ -8,7 +8,7 @@ import es.ediae.master.programacion.gestionusuario.service.UsuarioModel;
 
 public class UsuarioGetDTO {
 
-        private Long id;
+        private Integer id;
         private String nick_usuario;
         private String contrasena;
         private LocalDateTime fecha_hora_creacion;
@@ -22,7 +22,7 @@ public class UsuarioGetDTO {
         public UsuarioGetDTO() {
         }
 
-    public UsuarioGetDTO(String contrasena, LocalDateTime fecha_hora_creacion, LocalDate fecha_nacimiento, LocalTime hora_desayuno, Long id, String nick_usuario, String nombre, String primer_apellido, String segundo_apellido) {
+    public UsuarioGetDTO(String contrasena, LocalDateTime fecha_hora_creacion, LocalDate fecha_nacimiento, LocalTime hora_desayuno, Integer id, String nick_usuario, String nombre, String primer_apellido, String segundo_apellido) {
         this.contrasena = contrasena;
         this.fecha_hora_creacion = fecha_hora_creacion;
         this.fecha_nacimiento = fecha_nacimiento;
@@ -34,11 +34,11 @@ public class UsuarioGetDTO {
         this.segundo_apellido = segundo_apellido;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
