+++ src/main/java/es/ediae/master/programacion/gestionusuario/repository/IGeneroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import es.ediae.master.programacion.gestionusuario.entity.GeneroEntiity;
 
-public interface IGeneroRepository extends JpaRepository<GeneroEntiity, Long> {
+public interface IGeneroRepository extends JpaRepository<GeneroEntiity, Integer> {
 
 }
