+++ src/main/java/es/ediae/master/programacion/gestionusuario/repository/IUsuarioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import es.ediae.master.programacion.gestionusuario.entity.UsuarioEntity;
 
-public interface IUsuarioRepository extends JpaRepository<UsuarioEntity, Long> {
+public interface IUsuarioRepository extends JpaRepository<UsuarioEntity, Integer> {
 
 }
